+++ authorsite/contact/templates/contact/contact_page.html
@@ -15,6 +15,7 @@
                 {% endfor %}
             </div>
             {% endif %}
+        
         {% for field in form %}
         <div class="fieldWrapper form-group text-center justify-content-center">
             {% if field.errors %}
@@ -26,11 +27,20 @@
                 </div>
                 {% endfor %}
             {% endif %}
+
+            {% if field != form.spam_catcher %}
             <div class="row justify-content-center">
                 <div class="col-12 text-center">
                     <div class="mt-3 mb-0 text-uppercase h3 text-muted">{{ field.label_tag }}</div>
                 </div>
             </div>
+            {% else %}
+            <div class="row justify-content-center">
+                <div class="col-12 text-center">
+                    <div class="mt-3 mb-3 text-uppercase h3 text-muted">Spam Catcher: {{ settings.contact.SpamSettings.spam_question }}</div>
+                </div>
+            </div>
+            {% endif %}
 
             {% if field.help_text %}
             <div class="row">
@@ -49,6 +59,7 @@
             </div>
         </div>
         {% endfor %}
+        
         <div class="row justify-content-center mb-5">
          <button type="submit" value="send" class="btn btn-outline-warning p-2 newsletter-font text-uppercase shadow">Send My Message</button>
         </div>
